+++ packages/expect/src/__tests__/stacktrace.test.js
@@ -16,13 +16,22 @@ jestExpect.extend({
       pass: true,
     };
   },
+  toPreserveErrorCallStack(callback) {
+    try {
+      callback();
+    } catch (error) {
+      const assertionError = new jestExpect.JestAssertionError(error.message);
+      assertionError.stack = error.stack;
+      throw assertionError;
+    }
+  },
 });
 
 it('stack trace points to correct location when using matchers', () => {
   try {
     jestExpect(true).toBe(false);
   } catch (error) {
-    expect(error.stack).toContain('stacktrace.test.js:23');
+    expect(error.stack).toContain('stacktrace.test.js:32');
   }
 });
 
@@ -32,6 +41,22 @@ it('stack trace points to correct location when using nested matchers', () => {
       jestExpect(value).toBe(false);
     });
   } catch (error) {
-    expect(error.stack).toContain('stacktrace.test.js:32');
+    expect(error.stack).toContain('stacktrace.test.js:41');
+  }
+});
+
+it('stack trace points to correct location when throwing an instance of JestAssertionError', () => {
+  try {
+    jestExpect(() => {
+      const foo = () => bar();
+      const bar = () => baz();
+      const baz = () => {
+        throw new Error('Expected');
+      };
+
+      foo();
+    }).toPreserveErrorCallStack();
+  } catch (error) {
+    expect(error.stack).toContain('stacktrace.test.js:54');
   }
 });