+++ CHANGELOG.md
@@ -1,13 +1,15 @@
 ## master
 
-None for now
-
 ### Fixes
 
 * `[babel-jest]` moduleFileExtensions not passed to babel transformer.
   ([#4637](https://github.com/facebook/jest/issues/4637))
 
 ### Features
+
+* Expose `expect.JestAssertionError` to enable custom matchers to throw errors
+  with the call stack preserved.
+  ([#5138](https://github.com/facebook/jest/pull/5138))
 
 ### Chore & Maintenance